LayoutTests/fast/css/border-radius-property-value.html

<!DOCTYPE HTML>
<html>
<style>
 #element-container {
   border-radius:5px;
}
</style>
<div id="element-container"></div>
<script src="../js/resources/js-test-pre.js"></script>

<script>
if (window.testRunner)
    testRunner.dumpAsText();

description("Test for Bug: 80736 - CSSStyleDeclaration.getPropertyValue() for 'border-radius' returns null when value is set");

var element = document.getElementById("element-container");
var style = window.getMatchedCSSRules(element)[0].style;
shouldBe("style.getPropertyValue('border-radius')","'5px'");
shouldBe("style.getPropertyValue('border-top-left-radius')","'5px'");
shouldBe("style.getPropertyValue('border-top-right-radius')","'5px'");
shouldBe("style.getPropertyValue('border-bottom-left-radius')","'5px'");
shouldBe("style.getPropertyValue('border-bottom-right-radius')","'5px'");
</script>

<script src="../js/resources/js-test-post.js"></script>
</html>
