third_party/WebKit/LayoutTests/fast/css/border-radius-property-value-expected.txt

Test for Bug: 80736 - CSSStyleDeclaration.getPropertyValue() for 'border-radius' returns null when value is set

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS style.getPropertyValue('border-radius') is '5px'
PASS style.getPropertyValue('border-top-left-radius') is '5px'
PASS style.getPropertyValue('border-top-right-radius') is '5px'
PASS style.getPropertyValue('border-bottom-left-radius') is '5px'
PASS style.getPropertyValue('border-bottom-right-radius') is '5px'
PASS successfullyParsed is true

TEST COMPLETE
